requirements.txt
peft
transformers
accelerate
datasets
monai
torch
torchvision
pycocotools
numpy
matplotlib
Pillow
tqdm